Better separation of composegenertor v2 and v3
